+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/cap-config-writer
 
+## 0.6.4
+
+### Patch Changes
+
+-   Updated dependencies [0f3cf6b]
+    -   @sap-ux/project-access@1.25.1
+    -   @sap-ux/odata-service-inquirer@0.4.4
+
+## 0.6.3
+
+### Patch Changes
+
+-   Updated dependencies [f076dd3]
+    -   @sap-ux/project-access@1.25.0
+    -   @sap-ux/odata-service-inquirer@0.4.3
+
+## 0.6.2
+
+### Patch Changes
+
+-   Updated dependencies [0ae685e]
+    -   @sap-ux/project-access@1.24.0
+    -   @sap-ux/odata-service-inquirer@0.4.2
+
 ## 0.6.1
 
 ### Patch Changes