+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.3
+
+### Patch Changes
+
+-   Updated dependencies [0f3cf6b]
+    -   @sap-ux/project-access@1.25.1
+
+## 0.4.2
+
+### Patch Changes
+
+-   Updated dependencies [f076dd3]
+    -   @sap-ux/project-access@1.25.0
+
+## 0.4.1
+
+### Patch Changes
+
+-   Updated dependencies [0ae685e]
+    -   @sap-ux/project-access@1.24.0
+
 ## 0.4.0
 
 ### Minor Changes