+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/project-access
 
+## 1.25.1
+
+### Patch Changes
+
+-   0f3cf6b: feat: Add path to specification
+
+## 1.25.0
+
+### Minor Changes
+
+-   f076dd3: Add freestyle CAP app support.
+
+## 1.24.0
+
+### Minor Changes
+
+-   0ae685e: Add feature to cache node modules locally, consumed by specification
+
 ## 1.23.0
 
 ### Minor Changes